requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
Django==2.0.3
djangorestframework==3.7.7
idna==2.6
python-dateutil==2.7.0
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
